pyproject.toml
[tool.poetry]
name = "template"
version = "0.1.0"
description = "Template for nextcord discord bots"
authors = ["ooliver1 <oliverwilkes2006@icloud.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
nextcord = { extras = ["speed"], version = "^2.3.2" }
uvloop = "^0.17.0"
ooliver-botbase = { version = "^2.0.1", extras = ["db"] }
python-dotenv = "^0.21.0"
alembic = "^1.10.4"
taskipy = "^1.10.4"
delarva = { git = "https://github.com/ooliver1/delarva" }

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.3.0"
pre-commit = "^3.3.1"

[tool.taskipy.tasks]
start = "python -m botunderscorename"
lint = "pre-commit run --all-files"
precommit = "pre-commit install --install-hooks"

[tool.ruff]
select = [
    "E",
    "F",
    "W",
    "C90",
    "I",
    "N",
    "UP",
    "ANN",
    "S",
    "BLE",
    "B",
    "C4",
    "DTZ",
    "ISC",
    "G",
    "PIE",
    "Q",
    "RSE",
    "RET",
    "SIM",
    "TCH",
    "ARG",
    "PTH",
    "PGH",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "TRY",
    "RUF",
]
ignore = ["S101", "ANN101", "ANN102", "F405", "S311"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
